Adding a (re)binding for move-to-window-line-top-bottom
This used to be on M-r, but that's now gobbled by smartparens.  I also
have C-' and C-" (which also occasionally get clobbered) for moving to
the top and bottom, but if I get used to this binding (which has a
nice symmetry with C-l for recenter-top-bottom) I might delete those.