Add powershell mode for windows
It's pinned to melpa which I think should be the default repo, but it
looked necessary to be specific here; there's a few different packages
called "powershell", and they all look different across
melpa/melpa-stable/marmalade.  I mainly want the mode, but there's
also an inferior-mode which might be useful and this is the most
up-to-date looking one too.