Bugfix: interaction between quelpa-use-package and always-ensure
So I realised that use-package still wasn't using my quelpa fetcher,
and it turns out that if you have `use-package-always-ensure' enabled,
as I do, then it will always use melpa.  Alternatively, you can
configure it to always use quelpa -- but if you want the status-quo
melpa most of the time /except/ when there's a :quelpa block, you need
to activate some advice, for which quelpa provides a function.