Remove jsx-mode
Well, that's an innocuous comment: all this time I thought jsx-mode
was for Facebook's JSX templating, whereas in fact it was for a
new-dead statically-typed-JS experiment.  Emacs now has js-jsx-mode
built in too!

And the only reason I went looking was tracking down some bytecomp
warnings that keep opening up on startup that I was getting sick
of (for replace-string and beginning-of-buffer, about being intended
for interactive use only).  Both gone now.