Change modifier for windmove bindings
Bit of a hack; I don't use this as much as I used to (with the C-xo
hydra), but at the same time I never identified why it's necessary.
Org should conflict with it, with instructions to make them play
nicely together, but instead windmove clobbers org for some reason and
even playing with load ordering I wasn't able to change it.  This lets
me move on.