Bugfix: unpin magit from stable
This was causing issues with org-roam, because the version of
magit-section that was loaded was not the separately-packaged version,
but an ancient copy originally packaged with magit itself.

I'm not entirely sure I understand how ":pin" works, I thought it
would upgrade versions from melpa-stable, but it seems to pin the
version currently installed and not touch it.  I feel like I've been
bitten by this before.