other-window repeats: remove destructive operations
I've found it too confusing; a common example is opening calc (which
means you have at least 3 windows now), then C-xo out of there, and
typing the value from calc, which means if it starts with a 1 then you
delete the rest.. etc. Value derived minimal compared to the
paper-cuts.