Adding Bob Glickstein's unscroll package
Slightly tweaked; years ago I used and appreciated this, and then it
stopped working and I presumably didn't look too hard why.  Without
verifying, I'm assuming that C-v used to be bound to scroll-up, not
scroll-up-command (et al).  In essence, the core scroll commands are advised
to (conditionally) save state, but the condition checks on
last-command, which is the higher-level scroll-up-command, so /those/
are the ones we need to add the unscrollable property to.

(And "unscrollable", to indicate that something is scrollable and can
be undone, reminds me of Archer. "Well I learned inflammable and
flammable mean the same thing!")

Also tidied up a bit thanks to checkdoc.