Add consult-recent-xref
This is more or less what I was after with tags-view-mode, the first
substantial bit of elisp I wrote!  That was a bit more
featured (namely, I wanted to prune the history list so I could focus
on the important bits through a code path), but this is a fraction the
size utilising consult et al.