Big update: remove lsp-mode in favour of built-in eglot
Simplifies configuration a lot!  Big change for now is manual
invocation of 'eglot' (once per project, it figures it out for
mode+language combo after that), apart from clojure which is still
automatic (ie, interactive eglot vs hook-invoked eglot-ensure)

Also needed some configuration to enable "sub projects"; because eglot
uses the built-in project.el you need to customise any vc project
where the language root is a sub-directory --
project-vc-extra-root-markers does this, but then you can't use "C-x p
f" to find files /outside/ that logical language project.  The answer
is a dynamic "eglot-lsp-context" which is non-nil when eglot is using
it, so you can customise a project-find-functions to look for a
".eglot" cookie file instead